inital draft of the refactored/split search functionality
