+++ server/src/core/search/mod.rs
@@ -1,0 +1,2 @@
+mod search_executor;
+pub(crate) mod search_handler;
